refactor: ExceptionType 패키지 이동
