+++ src/main/java/lotto/exception/LottoException.java
@@ -1,7 +1,5 @@
 package lotto.exception;
 
-import lotto.domain.type.ExceptionType;
-
 public class LottoException extends RuntimeException {
     private ExceptionType type;
